refactor: split types & remove inline functions
