+++ src/components/Header/HeaderNav.tsx
@@ -8,13 +8,17 @@ type HeaderNavProps = {
 };
 
 export function HeaderNav({ className, onClose }: HeaderNavProps) {
+  const handleNavLinkStyles = ({ isActive }: { isActive: boolean }) => {
+    return isActive ? "gradient-text" : "";
+  };
+
   return (
     <nav className={`header-nav ${className}`}>
       <ul>
         {headerLinks.map(({ route, text }) => {
           return (
             <li key={route}>
-              <NavLink to={route} className={({ isActive }) => (isActive ? "gradient-text" : "")} onClick={onClose}>
+              <NavLink to={route} className={handleNavLinkStyles} onClick={onClose}>
                 {text}
               </NavLink>
             </li>
