+++ src/types/currencyType.ts
@@ -1,0 +1,6 @@
+export type CurrencyType = {
+  code: string;
+  name: string;
+  thumb: string;
+  decimal_digits: number;
+};
