+++ src/types/exchangeRate.ts
@@ -1,10 +1,3 @@
-export type CurrencyType = {
-  code: string;
-  name: string;
-  thumb: string;
-  decimal_digits: number;
-};
-
 export type ExchangeRateType = {
   code: string;
   value: number;
