+++ src/types/index.ts
@@ -1,5 +1,6 @@
 import { BankType } from "./banks";
 import { XOHLCType } from "./chart";
-import { CurrencyType, ExchangeRatesDataType, ExchangeRateType } from "./currencies";
+import { CurrencyType } from "./currencyType";
+import { ExchangeRatesDataType, ExchangeRateType } from "./exchangeRate"
 
 export type { BankType, CurrencyType, ExchangeRatesDataType, ExchangeRateType, XOHLCType };
